refactor: move usePopoverHover to a separate file
