+++ src/core/packages/chrome/navigation/src/components/popover/index.tsx
@@ -20,10 +20,11 @@ import React, {
 } from 'react';
 
 import { focusFirstElement } from '../../utils/focus_first_element';
-import { useClickToggle, useHover } from '../../hooks/use_mouse_management';
+import { useClickToggle } from '../../hooks/use_mouse_management';
 import { blurPopover } from './blur_popover';
 import { usePopoverOpen } from './use_popover_open';
 import { useKeyboardManagement } from './use_keyboard_management';
+import { usePopoverHover } from './use_popover_hover';
 
 const TOP_BAR_HEIGHT = 48;
 const TOP_BAR_POPOVER_GAP = 8;
@@ -67,7 +68,7 @@ export const SideNavPopover = ({
     clearClickOpened();
   }, [close, clearClickOpened]);
 
-  const { handleMouseEnter, handleMouseLeave, clearTimeout } = useHover(
+  const { handleMouseEnter, handleMouseLeave, clearTimeout } = usePopoverHover(
     persistent,
     isOpenedByClick,
     isSidePanelOpen,
