+++ src/core/packages/chrome/navigation/src/hooks/use_mouse_management.ts
@@ -9,10 +9,6 @@
 
 import { useCallback, useState } from 'react';
 
-import { useHoverTimeout } from './use_hover_timeout';
-
-const HOVER_DELAY = 100;
-
 /**
  * Hook for managing click-based opening (persistent mode)
  */
@@ -23,33 +19,4 @@ export const useClickToggle = () => {
   const clearClickOpened = useCallback(() => setIsOpenedByClick(false), []);
 
   return { isOpenedByClick, setClickOpened, clearClickOpened };
-};
-
-/**
- * Hook for mouse interactions
- */
-export const useHover = (
-  persistent: boolean,
-  isOpenedByClick: boolean,
-  isSidePanelOpen: boolean,
-  { open, close }: { open: () => void; close: () => void }
-) => {
-  const { setTimeout, clearTimeout } = useHoverTimeout();
-
-  const handleMouseEnter = useCallback(() => {
-    if (!persistent || !isOpenedByClick) {
-      clearTimeout();
-      if (!isSidePanelOpen) {
-        open();
-      }
-    }
-  }, [persistent, isOpenedByClick, isSidePanelOpen, clearTimeout, open]);
-
-  const handleMouseLeave = useCallback(() => {
-    if (!persistent || !isOpenedByClick) {
-      setTimeout(close, HOVER_DELAY);
-    }
-  }, [persistent, isOpenedByClick, setTimeout, close]);
-
-  return { handleMouseEnter, handleMouseLeave, clearTimeout };
 };
